services/text-detector/requirements.txt
fastapi
uvicorn[standard]
onnxruntime-gpu
numpy
pillow
python-multipart

